requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.6
django-heroku==0.3.1
django-seed==0.2.2
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.0.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
urllib3==1.26.3
whitenoise==5.2.0
